+++ nengo_bones/tests/test_generate_bones.py
@@ -49,6 +49,17 @@ def test_ci_scripts(tmpdir):
               - post command 1
           - template: docs
           - template: deploy
+          - template: remote-script
+            remote_script: test
+            output_name: remote-test
+            azure_name: azure-name
+            azure_group: azure-group
+            coverage: true
+          - template: remote-script
+            remote_script: docs
+            output_name: remote-docs
+            remote_vars:
+              remote_var: remote_val
     """,
     )
 
@@ -105,6 +116,17 @@ def test_ci_scripts(tmpdir):
     )
 
     assert has_line(".ci/deploy.sh", 'exe python -c "from dummy import version')
+
+    assert has_line(
+        ".ci/remote-test.sh",
+        "exe az vm start --resource-group azure-group --name azure-name",
+    )
+    assert has_line(".ci/remote-test.sh", "bash .ci/test.sh install")
+    assert has_line(".ci/remote-test.sh", 'COVERAGE_FILE="tmp/dummy')
+
+    assert has_line(".ci/remote-docs.sh", 'export remote_var="remote_val"')
+    assert has_line(".ci/remote-docs.sh", "bash .ci/docs.sh install")
+    assert has_line(".ci/remote-docs.sh", 'REMOTE_FAILED_FILE="tmp/dummy')
 
 
 def test_travis_yml(tmpdir):